k8s-fakeMetrics/requirements.txt
prometheus_client>=0.15,<0.20
python-dotenv>=1.0,<2.0
